+++ src/app/shared/index.ts
@@ -1,2 +1,2 @@
 export * from './enums/category.enum';
-export * from './models/product.model';
+export * from './models';