+++ src/app/shared/models/cart-item.model.ts
@@ -1,0 +1,7 @@
+export interface CartItem {
+    photo: string;
+    name: string;
+    price: number;
+    size: number;
+    quantity: number;
+}