+++ changelog.md
@@ -54,3 +54,13 @@ Task 3
 1-6. Створено методи та сервіси сервіси згідно з завданнями, застосовано їх в компоненті AboutComponent, a також в методі життєвого циклу onInit в консоль виведено значення.
 
 7. Директива застосована на блоці about-component__block (блок змінює колір при кліку на нього).
+
+
+Task 4
+
+1. Застосовано пайпи uppercase(first.component.html, product.component.html для модифікування назви товарів), titlecase(cart-list.component.html для модифікування опцій сортування), currency(cart-list.component.html, cart-item.component.html, first.component.html, product.component.html для відображення ціни), date(app.component.html для відображення дати і часу рефрешення сторінки).
+
+2. Змінено метод getProducts() сервису ProductService(повертає Observable). Застосовано async пайпу в product-list.component.html для відображення списку товарів в корзині.
+3-5. Створено OrderByPipe, яка приймає 2 параметри (shared/pipes/order-by.pipe.ts). Зареєстровано її в  SharedModule і експортовано. Застосовано в cart-list.component.html для сортування списку товарів по ціні, кількості, назві а також в зростаючому і спадному порядку.
+
+6. Експортовано із  SharedModule модулі CommonModule, FormsModule та видалено їх з інших модулів.